# api-docs/openapi.yaml
openapi: 3.0.0
info:
  title: API fuer das Leistungssystem REO
  version: '1.0.0-SNAPSHOT'
servers:
  - url: http://localhost:8443/rest/intern
security:
  - basicAuth: [ ]
paths:
  /personen:
    get:
      summary: 'Liefert zu Suchkriterien passende Personen'
      operationId: 'readPersonenBySearchParams'
      parameters:
        - name: vorname
          required: false
          in: query
          schema:
            type: string
        - name: nachname
          required: false
          in: query
          schema:
            type: string
        - name: iban
          required: false
          in: query
          schema:
            type: string
        - name: email
          required: false
          in: query
          schema:
            type: string     
      responses:
        '200':
          description: 'Alles erfolgreich'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Person'
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
    Anschrift:
      required: [ id ]
      properties:
        id:
          type: integer
          format: int64
        strasse:
          type: string
        hausnummer:
          type: string
        plz:
          type: string
        ort:
          type: string
        land:
          type: string
    Bankverbindung:
      required: [ id ]
      properties:
        id:
          type: integer
          format: int64
        iban:
          type: string
        bankname:
          type: string
        bankleitzahl:
          type: string
        bic:
          type: string
        kontonummer:
          type: string
        land:
          type: string
    Kommunikationsadresse:
      required: [ id ]
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        telefon:
          type: string      
    Person:
      required: [ id ]
      properties:
        id:
          type: integer
          format: int64
        vorname:
          type: string
        nachname:
          type: string
        geschlecht:
          enum:
            - u
            - m
            - w
            - x
            - d
        anschriften:
          type: array
          items:
            $ref: '#/components/schemas/Anschrift'
        kommunikationsadressen:
          type: array
          items:
            $ref: '#/components/schemas/Kommunikationsadresse'
        bankverbindungen:
          type: array
          items:
            $ref: '#/components/schemas/Bankverbindung'